beginning major restructuring of vboats code to improve code-reusability and read-ability, as well as establish a cleaner method for tuning processing params.
