+++ include/algorithms/vboats/vboats.h
@@ -2,93 +2,115 @@
 #define ROBOCOMMANDER_ALGORITHMS_VBOATS_H_
 
 #include <string>
-#include <chrono>
 #include <vector>
-#include <thread>
-#include <mutex>
 
-#include <opencv2/opencv.hpp>
-#include <opencv2/core/version.hpp> /* For OpenCV backwards compatibility */
-
+#include "base/definitions.h"
 #include "utilities/image_utils.h"
+#include "algorithms/vboats/umap_processing_params.h"
+#include "algorithms/vboats/vmap_processing_params.h"
+// #include "algorithms/vboats/image_processing_debugging.h"
 
 using namespace std;
 
-class Obstacle{
+typedef enum ContourFilterMethod{
+	PERIMETER_BASED = 1,
+	AREA_BASED = 2,
+}ContourFilterMethod;
+typedef enum UvMapFilterMethod{
+	STRIPPING_METHOD = 0,
+	SOBELIZED_METHOD = 1,
+}UvMapFilterMethod;
+typedef enum ImageAngleCorrectionType{
+	ROLL_CORRECTION = 0,
+	PITCH_CORRECTION = 1,
+	YAW_CORRECTION = 2,
+}ImageAngleCorrectionType;
+
+
+class Vboats{
 public:
-    Obstacle(vector<cv::Point> pts, vector<int> dBounds);
-    int dMin;
-    int dMax;
-    cv::Point maxXY;
-    cv::Point minXY;
-    double angle;
-    float distance;
-    cv::Point3f location;
-    void update(bool depth_based, float cam_baseline = 0, float cam_dscale = 0,
-         float* cam_focal = nullptr, float* cam_principal_point = nullptr,
-         float dtype_gain = 0, float aux_dist_factor = 0, bool verbose = true
-    );
-};
+	UmapProcessingParams umapParams;
+	VmapProcessingParams vmapParams;
 
-class VBOATS{
-private:
-     // Counters
-     int nObs = 0;
-     // Operational Flags
-     bool _is_ground_present = true;
-
-     // Debug Flags
-     bool _debug = false;
-     bool _flag_simulation = false;
-     bool _debug_contours = false;
-     bool _debug_windows = false;
-     bool _debug_obstacle_search = false;
 public:
      /** Constructors */
-     VBOATS();
-     ~VBOATS();
+     Vboats();
+     ~Vboats();
 
      /** Startup - Shutdown - Initialization Functions */
      void init();
      void update();
+	int process();
 
-     void filter_disparity_vmap(const cv::Mat& input, cv::Mat* output, vector<float>* thresholds, bool verbose = false, bool debug = false, bool visualize = false);
-     void filter_disparity_umap(const cv::Mat& input, cv::Mat* output, vector<float>* thresholds, bool verbose = false, bool debug = false, bool visualize = false);
+     cv::Mat remove_umap_deadzones(const cv::Mat& umap);
+     cv::Mat remove_vmap_deadzones(const cv::Mat& vmap);
 
-     bool is_ground_present();
-     int find_ground_lines(const cv::Mat& vmap, cv::Mat* rhos, cv::Mat* thetas, int hough_thresh = 100, bool verbose = false);
-     int find_ground_lines(const cv::Mat& vmap, cv::Mat* found_lines, int hough_thresh = 100, bool verbose = false);
-     int find_ground_lines(const cv::Mat& vmap, vector<cv::Vec2f>* found_lines, int hough_thresh = 100, bool verbose = false);
-     void get_hough_line_params(const float& rho, const float& theta, float* slope, int* intercept);
-     int estimate_ground_line(const vector<cv::Vec2f>& lines, float* best_slope, int* best_intercept, float* worst_slope, int* worst_intercept,
-          double gnd_deadzone = 2.0, double minDeg = 26.0, double maxDeg = 89.0, bool verbose = false, bool debug_timing = false
-     );
-     bool find_ground_line(const cv::Mat& vmap, float* best_slope, int* best_intercept,
-          double minDeg = 26.0, double maxDeg = 89.0, int hough_thresh = 100,
-          double gnd_deadzone = 2.0, bool verbose = false, bool debug_timing = false, bool visualize = false
-     );
+     void set_camera_info(cv::Mat K, float depth_scale, float baseline, bool verbose = false);
+     void set_camera_orientation(double roll, double pitch, double yaw, bool verbose = false);
+     void set_camera_orientation(double x, double y, double z, double w, bool verbose = false);
+     void set_camera_angle_offset(double offset_degrees);
+     void set_depth_denoising_kernel_size(int size);
 
-     void extract_contour_bounds(const vector<cv::Point>& contour, vector<int>* xbounds, vector<int>* dbounds, bool verbose = false);
-     void find_contours(const cv::Mat& umap, vector<vector<cv::Point>>* found_contours,
-          int filter_method = 1, float min_threshold = 30.0, int* offsets = nullptr,
-          float max_threshold = -1, bool verbose = false, bool visualize = false,
-          bool debug = false, bool debug_timing = false
-     );
+     void set_absolute_minimum_depth(float value);
+     void set_absolute_maximum_depth(float value);
+     void set_contour_filtering_method(std::string method);
+     void set_umap_processing_method(std::string method);
+     void set_vmap_processing_method(std::string method);
+     void set_image_angle_correction_type(std::string method);
 
-     int obstacle_search_disparity(const cv::Mat& vmap, const vector<int>& xLimits, vector<int>* yLimits,
-          int* pixel_thresholds = nullptr, int* window_size = nullptr,
-          std::vector<float> line_params = {}, vector<cv::Rect>* obs_windows = nullptr,
-          bool verbose = true, bool visualize = false, bool debug = false, bool debug_timing = false
-     );
-     int find_obstacles_disparity(const cv::Mat& vmap, const vector<vector<cv::Point>>& contours,
-           vector<Obstacle>* found_obstacles, std::vector<float> line_params, vector< vector<cv::Rect> >* obstacle_windows = nullptr,
-           bool verbose = false, bool debug_timing = false
-     );
+     void enable_angle_correction(bool flag = true);
+     void enable_filtered_depth_denoising(bool flag = true);
+     void enable_obstacle_data_extraction(bool flag = true);
 
-     int pipeline_disparity(const cv::Mat& disparity, const cv::Mat& umap, const cv::Mat& vmap,
-          vector<Obstacle>* obstacles, cv::Mat* uMorphElement = nullptr,
-          bool verbose = false, bool debug_timing = false, bool visualize = false
-     );
+private:
+     std::string classLbl = txt_bold_magenta() + "Vboats" + txt_reset_color();
+     // Counters
+     int _cam_info_count = 0;
+
+     // TODO: Section Name
+     float _hard_min_depth = 0.1;
+     float _hard_max_depth = 20.0;
+
+     // TODO: Section Name
+     float _fx                     = 0;
+     float _fy                     = 0;
+     float _px                     = 0;
+     float _py                     = 0;
+     float _baseline               = 0;
+     float _depth_scale            = 0;
+     float _Tx                     = 0;
+     float _depth2disparityFactor  = 0;
+
+     // TODO: Section Name
+     double _cam_roll    = 0.0;
+     double _cam_pitch   = 0.0;
+     double _cam_yaw     = 0.0;
+     double _cam_angle_offset     = 0.0;
+	int _filtered_depth_denoising_size = 2;
+
+     // TODO: Section Name
+     cv::Mat _cur_depth;
+     cv::Mat _cur_disparity;
+     cv::Mat _cur_umap;
+     cv::Mat _cur_vmap;
+
+     // TODO: Section Name
+     ContourFilterMethod _contourFiltMeth = PERIMETER_BASED;
+     UvMapFilterMethod _umapFiltMeth = SOBELIZED_METHOD;
+     UvMapFilterMethod _vmapFiltMeth = SOBELIZED_METHOD;
+     ImageAngleCorrectionType _angleCorrectionType = ROLL_CORRECTION;
+
+     // Operational Flags
+     bool _have_cam_info = false;
+     bool _is_ground_present = true;
+     bool _do_angle_correction = true;
+     bool _denoise_filtered_depth = true;
+     bool _do_obstacle_data_extraction = true;
+
+     // Debug Objects
+     bool _verbose = false;
+     UmapProcessingDebugObjects _umap_debugger;
+     VmapProcessingDebugObjects _vmap_debugger;
 };
 
 #endif // ROBOCOMMANDER_ALGORITHMS_VBOATS_H_
